Fixed various compilation bugs
